+++ lib/utils.dart
@@ -1,11 +1,12 @@
 library elec_server.utils;
 
-export 'src/utils/parse_custom_integer_range.dart';
+export 'src/utils/convert_xls_to_xlsx.dart';
 export 'src/utils/file_utils.dart';
 export 'src/utils/iterable_extensions.dart';
 export 'src/utils/layout_tree.dart';
 export 'src/utils/lib_plotly.dart';
+export 'src/utils/parse_custom_integer_range.dart';
+export 'src/utils/mailtrap.dart';
+export 'src/utils/string_extensions.dart';
 export 'src/utils/term_cache.dart';
 export 'src/utils/to_csv.dart';
-export 'src/utils/string_extensions.dart';
-export 'src/utils/convert_xls_to_xlsx.dart';