+++ test/db/lib_update_dbs_test.dart
@@ -41,9 +41,9 @@ Future<void> tests() async {
 
   ///---------------------------------------------------------------
   /// IESO
-  // var months = Month.utc(2024, 1).upTo(Month.utc(2024, 10));
+  // var months = Month.utc(2025, 1).upTo(Month.utc(2025, 1));
   // await updateIesoRtGenerationArchive(months: months);
-  // await updateIesoRtZonalDemandArchive(years: [2024]);
+  // await updateIesoRtZonalDemandArchive(years: [2025]);
 
   ///---------------------------------------------------------------
   /// ISONE
@@ -52,15 +52,18 @@ Future<void> tests() async {
   // await updateIsoneRtSystemLoad5minArchive(days: days, download: true);
 
   // await insertDays(DaLmpHourlyArchive(), days, gzip: true);
-  // final months = Month(2024, 6, location: IsoNewEngland.location)
-  //     .upTo(Month(2024, 6, location: IsoNewEngland.location));
+  final months = Month(2024, 7, location: IsoNewEngland.location)
+      .upTo(Month(2024, 7, location: IsoNewEngland.location));
   // await updateIsoneDaLmp(months: months, download: true);
+  // await updateIsoneMonthlyAssetNcpc(months: months, download: true);
+  // await updateIsoneDemandBids(months: months, download: false);
+  await updateIsoneRtEnergyOffers(months: months, download: false);
+
   // await updateIsoneRtLmp(months: months, download: true);
   // await updateIsoneRtLmp5Min(
   //     months: months, ptids: [4000], reportType: 'prelim', download: false);
   // await updateIsoneRtLmp5Min(
   //     months: months, ptids: [4000], reportType: 'final', download: false);
-
 
   // await updateIsoneZonalDemand([2021], download: false);
   // await updateIsoneZonalDemand(IntegerRange(2011, 2021));
@@ -79,7 +82,6 @@ Future<void> tests() async {
   // await updateCtSupplierBacklogRatesDb(months: months,
   //     externalDownload: true);
 
-
   // var years = IntegerRange(2020, 2024);
   // await updateCmpLoadArchive(years, setUp: true);
   // await updatePolygraphProjects(setUp: false);
@@ -90,7 +92,7 @@ Future<void> tests() async {
 
   ///------------------------------------------------------------------
   /// Weather
-  await insertNoaaTemperatures(download: true);
+  // await insertNoaaTemperatures(download: true);
   // await insertNormalTemperatures();
 }